+++ CHANGELOG.md
@@ -1,2 +1,12 @@
 # v1.0.0
+
 Document features and fixes
+
+# v1.0.1
+
+added plugin editor for app-configurator
+
+# v1.0.2
+
+- added translation of editor
+- added bar chart option